Minor updates and fixed some bugs, not fully functional.

+++ src/main/java/SOLW_V7/NearStructs.java
@@ -5,8 +5,8 @@ import kaptainwutax.mcutils.util.pos.BPos;
 public class NearStructs {
     BPos outpost;
     BPos treasure;
-    int dist;
-    public NearStructs(BPos a, BPos b, int c){
+    long dist;
+    public NearStructs(BPos a, BPos b, long c){
         this.outpost = a;
         this.treasure = b;
         this.dist = c;
@@ -15,5 +15,5 @@ public class NearStructs {
         return outpost;
     }
     public BPos getTreasure(){return treasure;}
-    public int getDist(){return dist;}
+    public long getDist(){return dist;}
 }
